requirements.txt
langgraph>=1.0.3
langchain-core>=1.2.0
langchain-openai>=1.1.0
fastapi>=0.122.0
uvicorn[standard]>=0.40.0
httpx>=0.28.0
langsmith>=0.1.0
python-dotenv>=1.0.0
